requirements.txt
streamlit
streamlit-mic-recorder
supabase
python-dotenv
requests